Remove block using the right button

+++ scripts/world.gd
@@ -6,7 +6,7 @@ const agent_left = preload("res://objects/agent_right.tscn")
 const door_scene = preload('res://objects/door.tscn') 
 const block_scene = preload('res://objects/block.tscn')
 
-# MAIN CLASSES
+# CLASSES
 class Agent:
 	var object
 	var last_update
@@ -14,7 +14,7 @@ class Agent:
 	var side
 	var influence_number
 
-	func _init(side, speed, influence_number=300):
+	func _init(side, influence_number=300):
 		if side == 'left':
 			self.object = agent_left.instance()
 			self.influence_number = influence_number
@@ -22,7 +22,7 @@ class Agent:
 			self.object = agent_right.instance()
 			self.influence_number = influence_number*(-1)
 		self.side = side
-		self.speed = speed
+		self.speed = 0.05
 		self.last_update = self.speed
 
 	func type():
@@ -63,7 +63,6 @@ class Door:
 	var timestamp
 	var object
 	var influence_number
-	var agent_speed
 	var born_speed
 
 	func _init(born_speed,influence_number=10000):
@@ -71,7 +70,6 @@ class Door:
 		self.timestamp =  born_speed + self.get_random_value()
 		self.born_speed = born_speed + self.get_random_value()
 		self.influence_number = influence_number
-		self.agent_speed = 0.02
 
 	func type():
 		return 'Door'
@@ -97,17 +95,17 @@ class Door:
 
 		if self.timestamp < 0.0:
 			if x == 0 and typeof(map.board[x+1][y]) == TYPE_INT and typeof(map.board[x+1][y-1]) == TYPE_INT and typeof(map.board[x+1][y+1]) == TYPE_INT:
-				var agent = Agent.new('left', self.agent_speed)
+				var agent = Agent.new('left')
 				map.set_object(agent, x+1, y)
 				self.timestamp = self.get_random_value() + self.born_speed
 			elif x == 39 and typeof(map.board[x-1][y]) == TYPE_INT and typeof(map.board[x-1][y-1]) == TYPE_INT and typeof(map.board[x-1][y+1]) == TYPE_INT:
-				var agent = Agent.new('right', self.agent_speed)
+				var agent = Agent.new('right')
 				map.set_object(agent, x-1, y)
 				self.timestamp = self.get_random_value() + self.born_speed
 
 	func get_random_value():
 		randomize()
-		return randf()/8
+		return randf()/20
 
 	func add_influence(map, x, y):
 		for tmp_x in range(map.width):
@@ -122,11 +120,10 @@ class Door:
 class Block:
 	var object
 	var influence_number
-	var side
-	func _init(influence_number=0):
+	
+	func _init(influence_number=1000):
 		self.object = block_scene.instance()
 		self.influence_number = influence_number
-		self.side = ''
 	
 	func type():
 		return 'Block'
@@ -159,7 +156,7 @@ class Map:
 	var object
 	
 	func _init(width, height, object):
-		self.update_time =1
+		self.update_time =0.5
 		self.last_update = self.update_time
 		self.tile_size = 20
 		self.width = width
@@ -205,25 +202,28 @@ class Map:
 	func create_doors():
 		# right door
 		for y in [2,3,4,14,15,16,23,24,25]:
-			var door = Door.new(0.005, 10000)
+			var door = Door.new(0.1, 10000)
 			self.set_object(door, 0, y)
-		# left_door
-		for y in [2,3,4,14,15,16,23,24,25]:
 			var door = Door.new(0.1,-10000)
 			self.set_object(door, 39, y)
 
 	func set_object(instance, x, y):
+		if typeof(self.board[x][y]) != TYPE_INT:
+			return
+
 		self.object.add_child(instance.object)
 		self.board[x][y] = instance
 		instance.add_influence(self, x, y)
 		instance.object.set_pos(Vector2(self.tile_size*x, self.tile_size*y))
-	
+
 	func remove_object(x, y):
 		var instance = self.board[x][y]
+		if typeof(instance) == TYPE_INT:
+			return
 		instance.remove_influence(self, x, y)
 		self.board[x][y] = 0
 		instance.delete(self)
-		
+
 	func move_object(instance, x, y, next_x, next_y):
 		instance.remove_influence(self, x, y)
 		instance.add_influence(self, next_x, next_y)
@@ -238,7 +238,6 @@ class Map:
 		self.last_update -= delta
 		if self.last_update > 0:
 			return
-
 		self.last_update = self.update_time
 		for x in range(self.width):
 			for y in range(self.height):
@@ -249,18 +248,7 @@ class Map:
 	func update_door_speed(left_door_speed, right_door_speed):
 		for y in [2,3,4,14,15,16,23,24,25]:# left_door
 			self.board[0][y].born_speed = left_door_speed
-		for y in [2,3,4,14,15,16,23,24,25]:# right_door
 			self.board[39][y].born_speed = right_door_speed
-	
-	func update_agent_speed(speed):
-		for x in range(self.width):
-			for y in range(self.height):
-				var obj = self.board[x][y]
-				if typeof(obj) != TYPE_INT:
-					if obj.type() == 'Agent':
-						obj.speed = speed
-					if obj.type() == 'Door':
-						obj.agent_speed = speed
 
 class Game:
 	var menu_items
@@ -268,8 +256,7 @@ class Game:
 	
 	func _init(menu, map):
 		self.map = map
-		
-		var items_name = ['start_button','stop_button', 'agent_speed','left_door_speed',
+		var items_name = ['start_button','stop_button','left_door_speed',
 						  'right_door_speed', 'status','reset_button_agents', 'reset_train_station']
 		self.menu_items = {}
 		for name in items_name:
@@ -289,8 +276,7 @@ class Game:
 	func update(delta):
 		if self.menu_items['start_button'].is_pressed():
 			self.menu_items['status'].set_text('Running')
-			self.map.update_agent_speed(self.get_speed('agent_speed'))
-			self.map.update_door_speed( self.get_speed('left_door_speed'), self.get_speed('right_door_speed'))
+			self.map.update_door_speed(self.get_speed('left_door_speed'), self.get_speed('right_door_speed'))
 		if self.menu_items['stop_button'].is_pressed():
 			self.menu_items['status'].set_text('Stopped')
 		if self.menu_items['reset_button_agents'].is_pressed():
@@ -299,38 +285,47 @@ class Game:
 		if self.menu_items['reset_train_station'].is_pressed():
 			self.map.clear_blocks()
 			self.menu_items['status'].set_text('Stopped')
-			
 		self.map.update(delta, self.is_stopped())
 
 	func get_speed(button_name):
+		# RETURN SPEED BASED ON BUTTON SELECTION
+		# SLOW (0.5) , NORMAL (0.3), FAST (0.1)
 		var speed_index = self.menu_items[button_name].get_selected()
 		var speed = 0
-		if  speed_index == 0: # slow
+		if  speed_index == 0:
+			speed = 0.5
+		elif speed_index == 2:
 			speed = 0.1
-		elif speed_index == 2: # fast
-			speed = 0.02
 		else: # normal
-			speed = 0.035
+			speed = 0.3
 		return speed
 
 #  GAME LOOP
 var game
 
 func _ready():
+	# SETUP THE GAME 
 	set_process_input(true)
 	set_process(true)
 	var map = Map.new(40,30, get_node('.'))
 	game = Game.new(get_node('/root/world/menu'), map)
 
 func _process(delta):
+	# UPDATE THE GAME
 	game.update(delta)
 
 func _input(ev):
-	if (ev.type==InputEvent.MOUSE_BUTTON):
-		if ev.button_index == BUTTON_LEFT and ev.pressed and game.is_stopped() == "Stopped":
-			var block = Block.new()
+	# GET INPUT EVENT
+	if (ev.type==InputEvent.MOUSE_BUTTON): # MOUSE EVENT
+		if ev.pressed and game.is_stopped() == "Stopped":
+			# CALCULATE THE BOARD POSITION
 			var x = int(ev.pos.x) / game.map.tile_size
 			var y = int(ev.pos.y) / game.map.tile_size
-			if x > game.map.width or y > game.map.height:
+			# VERIFY THE BOARD LIMIT
+			if x > game.map.width-2 or y > game.map.height-2:
 				return
-			game.map.set_object(block, x, y)
+			if ev.button_index == BUTTON_LEFT:  # CREATE OBJECT WHEN BUTTON LEFT IS PRESSED
+				var block = Block.new()
+				game.map.set_object(block, x, y)
+			elif ev.button_index == BUTTON_RIGHT: # REMOVE OBJECT WHEN BUTTON LEFT IS PRESSED
+					game.map.remove_object(x,y)
